Ramayana4Everyone-main/chatbot/requirements.txt
gunicorn
python-dotenv
langchain
langchain_google_genai
flask
flask_cors